Convert EventConditionForm to functional TS component
